+++ src/utils/upload_utils.ts
@@ -6,7 +6,7 @@ import { API_ROUTES, HttpStatusCodes } from "../constants";
 import { PlanLimitsHandler } from './pricing_utils';
 import { formatDatetime, readFile } from './common';
 import { s3UploaderUtils } from '../init/s3_uploader';
-import { RepoPlanLimitsState } from './repo_utils';
+import { RepoPlanLimitsState } from './repo_state_utils';
 
 
 export const uploadRepoToServer = async (accessToken: string, data: any, repoId=null) => {
